+++ mocapi-autoconfigure/src/test/java/com/callibrity/mocapi/transport/http/StreamableHttpAutoConfigurationTest.java
@@ -110,7 +110,8 @@ class StreamableHttpAutoConfigurationTest {
           new McpRequestValidator(List.of("localhost")),
           new McpHeaderValidator(),
           objectMapper,
-          ContextSnapshotFactory.builder().build());
+          ContextSnapshotFactory.builder().build(),
+          0L);
     }
   }