+++ mocapi-streamable-http-transport/src/test/java/com/callibrity/mocapi/transport/http/StreamableHttpControllerTest.java
@@ -75,7 +75,8 @@ class StreamableHttpControllerTest {
             new McpRequestValidator(List.of("localhost")),
             new McpHeaderValidator(),
             objectMapper,
-            ContextSnapshotFactory.builder().build());
+            ContextSnapshotFactory.builder().build(),
+            0L);
   }
 
   // --- helpers ---